frametime type change
비주얼스투디오님의 충고를 받아 더블로 변경

+++ Run_Time_Framework.cpp
@@ -311,7 +311,7 @@ GLvoid CRun_time_Framework::Update() {
 		break;
 	}
 
-	if (FRAMETIME > 1000.0 / FPS_TIME) {
+	if ((double)FRAMETIME > 1000.0 / FPS_TIME) {
 		m_pCurrScene->Update(FRAMETIME);
 
 		Prevtime = current_time;
